+++ microbench/bfs/include/pando-bfs-galois/sssp.hpp
@@ -296,7 +296,7 @@ pando::Status SSSPMDLCSR(G& graph, std::uint64_t src, HostLocalStorage<MDWorkLis
         if (found) {
           graph.setDataOnly(srcID, 0);
 
-          std::int16_t srcHost = graph.getPhysicalHostID(src);
+          std::int64_t srcHost = graph.getPhysicalHostID(src);
           if (srcHost == pando::getCurrentPlace().node.id) {
             PANDO_CHECK(fmap(toRead[0], pushBack, srcID));
           }
