+++ pando-rt/examples/multi_pxns/onecore_scatter_samedim_arrays_values.cpp
@@ -32,7 +32,7 @@ void setValue(pando::GlobalPtr<std::int64_t> localArray, std::int64_t value,
 void scatterValues(pando::GlobalPtr<pando::GlobalPtr<std::int64_t>>& distArray,
                    std::int32_t numCoresPerNode, pando::GlobalPtr<bool> done) {
   const auto placeDims = pando::getPlaceDims();
-  std::int16_t thisNodeId = pando::getCurrentPlace().node.id;
+  std::int64_t thisNodeId = pando::getCurrentPlace().node.id;
   assert(thisNodeId == 0);
 
   // Initialize values.
@@ -44,7 +44,7 @@ void scatterValues(pando::GlobalPtr<pando::GlobalPtr<std::int64_t>>& distArray,
   done[thisNodeId] = true;
 
   // Scatter values to remote PXNs.
-  for (std::int16_t ipxn = 0; ipxn < placeDims.node.id; ++ipxn) {
+  for (std::int64_t ipxn = 0; ipxn < placeDims.node.id; ++ipxn) {
     pando::Place destPlace{pando::NodeIndex{ipxn}, pando::anyPod, pando::anyCore};
     pando::GlobalPtr<std::int64_t> remoteArray = distArray[ipxn];
     PANDO_CHECK(pando::executeOn(destPlace, &setValue, remoteArray, solution, numCoresPerNode,
@@ -60,7 +60,7 @@ void correctnessCheck(pando::GlobalPtr<pando::GlobalPtr<std::int64_t>> output,
   const auto placeDims = pando::getPlaceDims();
   // Check correctness.
   bool check_correctness{true};
-  for (std::int16_t ipxn = 0; ipxn < placeDims.node.id; ++ipxn) {
+  for (std::int64_t ipxn = 0; ipxn < placeDims.node.id; ++ipxn) {
     pando::GlobalPtr<std::int64_t> localArray = output[ipxn];
     for (std::int32_t c = 0; c < numCoresPerNode; ++c) {
       if (localArray[c] != solution) {
@@ -78,8 +78,8 @@ void correctnessCheck(pando::GlobalPtr<pando::GlobalPtr<std::int64_t>> output,
 /**
  * @brief Wait until invoked tasks complete.
  */
-void waitUntil(pando::GlobalPtr<bool> dones, std::int16_t numNodes) {
-  for (std::int16_t n = 0; n < numNodes; ++n) {
+void waitUntil(pando::GlobalPtr<bool> dones, std::int64_t numNodes) {
+  for (std::int64_t n = 0; n < numNodes; ++n) {
     pando::waitUntil([dones, n]() {
       return dones[n];
     });
@@ -105,7 +105,7 @@ int pandoMain(int, char**) {
   const auto thisPlace = pando::getCurrentPlace();
 
   if (thisPlace.node.id == 0) {
-    std::int16_t numPXNs = placeDims.node.id;
+    std::int64_t numPXNs = placeDims.node.id;
     std::int8_t numPodsPerPXNs = placeDims.pod.x * placeDims.pod.y;
     std::int32_t numCoresPerNode = placeDims.core.x * placeDims.core.y * numPodsPerPXNs;
     auto mmResource = pando::getDefaultMainMemoryResource();
@@ -118,7 +118,7 @@ int pandoMain(int, char**) {
     pando::GlobalPtr<bool> dones =
         static_cast<pando::GlobalPtr<bool>>(mmResource->allocate(sizeof(bool) * numPXNs));
     // Remote and local memory allocation
-    for (std::int16_t n = 0; n < placeDims.node.id; ++n) {
+    for (std::int64_t n = 0; n < placeDims.node.id; ++n) {
       const pando::Place otherPlace{pando::NodeIndex{n}, pando::anyPod, pando::anyCore};
       auto result =
           pando::allocateMemory<std::int64_t>(numCoresPerNode, otherPlace, pando::MemoryType::Main);
@@ -135,7 +135,7 @@ int pandoMain(int, char**) {
     correctnessCheck(distArray, numCoresPerNode);
 
     // Deallocate arrays
-    for (std::int16_t n = 0; n < placeDims.node.id; ++n) {
+    for (std::int64_t n = 0; n < placeDims.node.id; ++n) {
       pando::GlobalPtr<std::int64_t> localArray = distArray[n];
       pando::deallocateMemory(localArray, sizeof(std::int64_t) * numCoresPerNode);
     }
